Update button sizes and disable old launch script
The app no longer opens System Prefernces on launch.

Since the Control Strip will now sit in the corner of the
Touch Bar, the sizes of the buttons needed to be tweaked in
order to show all of them. Unfortunately, the Control Strip
arrow icon eats up a little more screen space than I'd like,
but there doesn't seem to be anything I can do about it.