Disable App Sandboxing
Due to tightening App Sandboxing restrictions in macOS Mojave and Catalina, this
is needed in order for a script packaged with the app to kill another process (the
Control Strip).

See: https://stackoverflow.com/questions/49544566/operation-not-permitted-when-executing-killall-with-swift

Note: this app is already using private frameworks and is not distributed via the
App Store.